CRD confirmed working with Postman :)

+++ Java EE/src/main/java/com/qa/business/service/Servicable.java
@@ -6,8 +6,8 @@ public interface Servicable {
 
 	String add(String entity);
 
-	String delete(int id);
+	String delete(Long id);
 
-	String get(int id);
+	String get(Long id);
 
 }
