+++ Java EE/src/main/java/com/qa/rest/Pointable.java
@@ -6,8 +6,8 @@ public interface Pointable {
 
 	String add(String game);
 
-	String delete(int id);
+	String delete(Long id);
 
-	String get(int id);
+	String get(Long id);
 
 }
